requirements.txt
langchain
langchain-community
streamlit
openai
python-dotenv
pymongo
